k3s-lap-data-spark-app/requirements-dev.txt
pyspark==3.5.1
pydantic
black
flake8
isort
python-dotenv